backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
django-pandas==0.6.2
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
numpy==1.19.3
pandas==1.1.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
